// frontend/src/app/api/models/dashboard-history.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { HistorySeries } from '../models/history-series';
export interface DashboardHistory {
  range: '1h' | '6h' | '24h' | '7d' | '30d';
  resolution: 'raw' | 'minute' | 'hour';
  series: Array<HistorySeries>;
  since: string;
  stepS: number;
  ts: Array<string>;
  until: string;
}